pyproject.toml
[project]
name = "vision-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.3.0",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "matplotlib>=3.9.0",
    "mcp>=1.0.0",
    "medmnist>=3.0.2",
    "numpy>=2.3.4",
    "pandas>=2.2.0",
    "pillow>=12.0.0",
    "pydantic>=2.10.0",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "python-multipart>=0.0.12",
    "scikit-learn>=1.4.0",
    "seaborn>=0.13.0",
    "sentence-transformers>=5.1.2",
    "transformers>=4.40.0",
    "datasets>=2.13.0",
    "umap-learn>=0.5.3",
    "uvicorn>=0.34.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
include = ["vision_rag*"]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.mypy]
python_version = "3.12"
files = ["vision_rag"]
ignore_missing_imports = true
disallow_untyped_defs = false
pretty = true
no_implicit_optional = false
